add interrupt handler parameter
