+++ src/rust/src/parport/irq.rs
@@ -25,12 +25,12 @@ impl Irq {
         }
     }
 
-    pub unsafe fn register(&mut self, parport: Arc<IRQLock<Parport>>) -> Result {
+    pub unsafe fn register(&mut self, parport: Arc<IRQLock<Parport>>, handler: unsafe extern "C" fn(*mut bindings::excp_entry_t, bindings::excp_vec_t, *mut c_void) -> c_int) -> Result {
         if self.registered {
             return Err(-1);
         }
 
-        let handler = interrupt_handler;
+        //let handler = interrupt_handler;
         self.arc_ptr = Arc::into_raw(parport);
         let result = unsafe {
             bindings::register_irq_handler(
@@ -63,31 +63,4 @@ impl Drop for Irq {
             }
         }
     }
-}
-
-unsafe fn deref_locked_state<'a>(state: *mut c_void) -> &'a IRQLock<Parport> {
-    // caller must guarantee `state`, and the object it points to, was not mutated
-    //
-    // caller must not drop the strong reference count of the containing `Arc` to 0 while
-    // the returned reference exists
-    let l = state as *const IRQLock<Parport>;
-    unsafe { l.as_ref() }.unwrap()
-}
-
-pub unsafe extern "C" fn interrupt_handler(
-    _excp: *mut bindings::excp_entry_t,
-    _vec: bindings::excp_vec_t,
-    state: *mut c_void,
-) -> c_int {
-    let p = unsafe { deref_locked_state(state) };
-    let mut l = p.lock();
-    l.set_ready();
-
-    // IRQ_HANDLER_END
-    unsafe {
-        bindings::apic_do_eoi();
-    }
-    0
-    // l falls out of scope here, releasing the lock and reenabling interrupts after
-    // IRQ_HANDLER_END. Redundant, but should work correctly.
 }
