+++ src/rust/src/parport/mod.rs
@@ -1,7 +1,12 @@
 use bitfield::bitfield;
+use core::{
+    ffi::{c_int, c_void},
+    ptr::null,
+};
 
 use crate::prelude::*;
 
+use crate::kernel::bindings;
 use chardev::NkCharDev;
 use irq::Irq;
 use portio::ParportIO;
@@ -69,7 +74,7 @@ impl Parport {
         {
             let mut locked_p = shared_p.lock();
             unsafe {
-                locked_p.irq.register(shared_p.clone()).inspect_err(|e| {
+                locked_p.irq.register(shared_p.clone(), interrupt_handler).inspect_err(|e| {
                     error!("Failed to register interrupt handler. Error code {e}.")
                 })?;
             }
@@ -209,3 +214,31 @@ register_shell_command!("parport", "parport", |_, _| {
         .inspect_err(|_| vc_println!("Unable to bring up parport device!"))
         .as_error_code()
 });
+
+
+unsafe fn deref_locked_state<'a>(state: *mut c_void) -> &'a IRQLock<Parport> {
+    // caller must guarantee `state`, and the object it points to, was not mutated
+    //
+    // caller must not drop the strong reference count of the containing `Arc` to 0 while
+    // the returned reference exists
+    let l = state as *const IRQLock<Parport>;
+    unsafe { l.as_ref() }.unwrap()
+}
+
+pub unsafe extern "C" fn interrupt_handler(
+    _excp: *mut bindings::excp_entry_t,
+    _vec: bindings::excp_vec_t,
+    state: *mut c_void,
+) -> c_int {
+    let p = unsafe { deref_locked_state(state) };
+    let mut l = p.lock();
+    l.set_ready();
+
+    // IRQ_HANDLER_END
+    unsafe {
+        bindings::apic_do_eoi();
+    }
+    0
+    // l falls out of scope here, releasing the lock and reenabling interrupts after
+    // IRQ_HANDLER_END. Redundant, but should work correctly.
+}
